+++ CHANGELOG.md
@@ -6,6 +6,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 
 ## Fixed/enhanced
 
+- Support for unordered matching which brings ~3x throughput improvements! Just pass `--no-preserve-order` flag.
 - `--only-analyze` with `--no-preserve-order` prevent OOM
 - `--only-analyze` option `--queue-size` to specify the Java executor service queue size
 - Flag `--show-analysis-components` to print out a list of available text analysis components